pyproject.toml
[tool.poetry]
name = "genshin"
version = "0.1.0"
description = ""
authors = ["nishiharu <checompetition@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
transformers = "^4.36.2"
torch = "^2.1.2"
loguru = "^0.7.2"
uvicorn = "^0.25.0"
fastapi = "^0.108.0"
streamlit = "^1.29.0"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.2"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
